+++ services/src/main/java/org/keycloak/services/resources/admin/SAMLFederationResource.java
@@ -149,6 +149,21 @@ public class SAMLFederationResource {
         return representation;
     }
 
+    @POST
+    @Path("instances/{id}/refresh")
+    public void resfreshFederation(@PathParam("id") String internalId) {
+        this.auth.realm().requireManageIdentityProviders();
+        //create alias representation
+        FederationModel model = realm.getSAMLFederationById(internalId);
+        if (model == null) {
+            throw new NotFoundException();
+        }
+
+        FederationProvider federationProvider = SAMLFederationProviderFactory.getSAMLFederationProviderFactoryById(session, model.getProviderId()).create(session,model,this.realm.getId());
+        federationProvider.updateSamlEntities();
+
+    }
+
     /**
      * Delete the identity provider federation, along with all its IdPs
      *